Totally remove rabbitmq, rename some stuff to avoid namespace issues
