+++ aleph/settings.py
@@ -138,16 +138,5 @@ INDEX_READ = env.to_list('ALEPH_INDEX_READ', [INDEX_WRITE])
 
 # Disable delayed processing via queue
 EAGER = env.to_bool('ALEPH_EAGER', DEBUG)
-QUEUE_PREFIX = env.get('ALEPH_QUEUE_PREFIX', APP_NAME)
-QUEUE_NAME = '%s_worker' % QUEUE_PREFIX
-QUEUE_ROUTING_KEY = 'worker.process'
-
-BROKER_URI = 'amqp://guest:guest@localhost:5672//'
-BROKER_URI = env.get('ALEPH_BROKER_URI', BROKER_URI)
 
 sls.REDIS_URL = sls.REDIS_URL or 'redis://redis:6379/0'
-
-STACKDRIVER_TRACE_PROJECT_ID = env.get('ALEPH_STACKDRIVER_TRACE_PROJECT_ID')
-TRACE_SAMPLING_RATE = float(env.get('ALEPH_TRACE_SAMPLING_RATE', 0.10))
-CELERY_TRACE_SAMPLING_RATE = env.get('ALEPH_CELERY_TRACE_SAMPLING_RATE', 0.01)
-CELERY_TRACE_SAMPLING_RATE = float(CELERY_TRACE_SAMPLING_RATE)
