+++ requirements-toolkit.txt
@@ -5,8 +5,8 @@ urlnormalizer==1.2.3
 followthemoney==1.13.1
 followthemoney-util==1.13.1
 fingerprints==0.6.5
-servicelayer[google]==1.2.2
-servicelayer[amazon]==1.2.2
+servicelayer[google]==1.3.0
+servicelayer[amazon]==1.3.0
 normality==1.0.0
 pantomime==0.3.3
 alephclient==1.1.5
